Make output more similar to rpm by changing "Description" to "Summary" in output.  Maybe at some point the extended description from metadata.xml will be stored in portage and can be extracted then by epm for Description. Also changed "Build Host" from CHOST to HOSTNAME.
Modified to read environment.bz2 instead of the ebuild for some values.
This is more reliable because environment.bz2 is the state after ebuild
and eclass processing.  Epm was confused by ebuilds that have the HOMEPAGE
and DESCRIPTION in an eclass. (eg. xmms-mpg123)

Fixed up display of values that contain double and single quotes
(eg. libmad)

Added error checking for open()